Move I18n-formatter unit tests
Fixes 1796.
